Fix security vulnerability that echoed out all ENV vars.
In a mistaken attempt to make defining ENV vars easier, Elixir
would define every ENV var using Webpack's `DefinePlugin`.
This was a mistake.  The entire ENV of the server generating the
bundle would be serialized into an object and delivered in the
JavaScript bundle.  This code has been removed.  You need to allow
ENV variables using the webpack's `DefinePlugin` or `EnvironmentPlugin`.
To make this easier, Elixir now has a `mix.env` method that forwards
all arguments on to `EnvironmentPlugin`.